+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
-## July 30, 2021
+## June 30, 2021
+
+### `@datorama/akita-ng-effects (v1.0.3)`:
+
+- `hotfix`: Effects should be subscribed only once
 
 ### `@datorama/akita-ng-effects (v1.0.2)`:
 
 - Effects should work with multiple eager modules
 
-## July 7, 2021
+## June 7, 2021
 
 ### `@datorama/ng-entity-service (v6.1.0)`: